feat: delete question

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -4,6 +4,7 @@ import { QuestionsRepository } from '../repositories/questions-repository';
 
 interface DeleteQuestionUseCaseRequest {
 	questionId: string;
+	authorId: string;
 }
 
 type DeleteQuestionUseCaseResponse = Record<string, never>;
@@ -12,12 +13,18 @@ export class DeleteQuestionUseCase {
 	constructor(private questionsRepository: QuestionsRepository) {}
 
 	async execute(
-		data: DeleteQuestionUseCaseRequest,
+		request: DeleteQuestionUseCaseRequest,
 	): Promise<DeleteQuestionUseCaseResponse> {
-		const question = await this.questionsRepository.findById(data.questionId);
+		const question = await this.questionsRepository.findById(
+			request.questionId,
+		);
 
 		if (!question) {
 			throw new Error('Question not found.');
+		}
+
+		if (request.authorId !== question.authorId.toValue()) {
+			throw new Error('Not allowed.');
 		}
 
 		await this.questionsRepository.delete(question);
